requirements.txt
kafka-python==2.0.2
dash==1.16.3
